# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  device.Device:
    properties:
      created_at:
        type: string
      firmware_version:
        type: string
      id:
        type: string
      last_seen:
        type: string
      metadata:
        type: object
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  device.ProvisionResponse:
    properties:
      device_uuid:
        type: string
      heartbeat_interval_ms:
        type: integer
      mqtt_broker_url:
        type: string
      mqtt_command_topic:
        type: string
      mqtt_telemetry_topic:
        type: string
      sampling_interval_ms:
        type: integer
    type: object
...
  dto.CreateDeviceRequest:
    properties:
      firmware_version:
        type: string
      metadata:
        type: object
      name:
        type: string
      status:
        type: string
      type:
        type: string
    required:
    - name
    - type
    type: object
  dto.CreateRuleRequest:
    properties:
      enabled:
        type: boolean
      metric:
        type: string
      name:
        type: string
      operator:
        type: string
      threshold:
        type: number
    required:
    - metric
    - name
    - operator
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  dto.HeartbeatRequest:
    properties:
      status:
        type: string
    type: object
...
  dto.ProvisionDeviceRequest:
    properties:
      capabilities:
        type: object
      device_type:
        type: string
      firmware_version:
        type: string
      hardware_id:
        type: string
    required:
    - device_type
    - hardware_id
    type: object
...
  dto.UpdateDeviceRequest:
    properties:
      firmware_version:
        type: string
      metadata:
        type: object
      name:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  dto.UpdateRuleRequest:
    properties:
      enabled:
        type: boolean
      metric:
        type: string
      name:
        type: string
      operator:
        type: string
      threshold:
        type: number
    type: object
...
  ota.FirmwareArtifact:
    properties:
      checksum_sha256:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      filename:
        type: string
      id:
        type: string
      object_key:
        type: string
      size_bytes:
        type: integer
      version:
        type: string
    type: object
...
  rule.Alert:
    properties:
      created_at:
        type: string
      device_id:
        type: string
      id:
        type: string
      message:
        type: string
      metric:
        type: string
      observed_value:
        type: number
      operator:
        type: string
      rule_id:
        type: string
      telemetry_id:
        type: string
      threshold:
        type: number
    type: object
  rule.Rule:
    properties:
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      metric:
        type: string
      name:
        type: string
      operator:
        type: string
      threshold:
        type: number
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: Distributed fleet monitoring and control API.
  title: ARGUS API
  version: "1.0"
paths:
  /alerts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rule.Alert'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List rule alerts
      tags:
      - rules
  /devices:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/device.Device'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List devices
      tags:
      - devices
    post:
      consumes:
      - application/json
      parameters:
      - description: Device payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/device.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create device
      tags:
      - devices
  /devices/{deviceID}:
    delete:
      parameters:
      - description: Device ID
        in: path
        name: deviceID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete device
      tags:
      - devices
    get:
      parameters:
      - description: Device ID
        in: path
        name: deviceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/device.Device'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get device
      tags:
      - devices
    put:
      consumes:
      - application/json
      parameters:
      - description: Device ID
        in: path
        name: deviceID
        required: true
        type: string
      - description: Device update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/device.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update device
      tags:
      - devices
...
  /devices/{deviceID}/heartbeat:
    post:
      consumes:
      - application/json
      parameters:
      - description: Device ID
        in: path
        name: deviceID
        required: true
        type: string
      - description: Heartbeat payload
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.HeartbeatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/device.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Record device heartbeat
      tags:
      - devices
...
  /ota/firmware:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ota.FirmwareArtifact'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List firmware artifacts
      tags:
      - ota
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Firmware version
        in: formData
        name: version
        required: true
        type: string
      - description: Firmware binary
        in: formData
        name: firmware
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ota.FirmwareArtifact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Upload firmware artifact
      tags:
      - ota
  /ota/firmware/{firmwareID}:
    get:
      parameters:
      - description: Firmware artifact ID
        in: path
        name: firmwareID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ota.FirmwareArtifact'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get firmware artifact
      tags:
      - ota
  /provision:
    post:
      consumes:
      - application/json
      parameters:
      - description: Provisioning payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ProvisionDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/device.ProvisionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Provision device
      tags:
      - devices
  /rules:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rule.Rule'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List telemetry rules
      tags:
      - rules
    post:
      consumes:
      - application/json
      parameters:
      - description: Rule payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/rule.Rule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create telemetry rule
      tags:
      - rules
  /rules/{ruleID}:
    delete:
      parameters:
      - description: Rule ID
        in: path
        name: ruleID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete telemetry rule
      tags:
      - rules
    get:
      parameters:
      - description: Rule ID
        in: path
        name: ruleID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rule.Rule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get telemetry rule
      tags:
      - rules
    put:
      consumes:
      - application/json
      parameters:
      - description: Rule ID
        in: path
        name: ruleID
        required: true
        type: string
      - description: Rule update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rule.Rule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update telemetry rule
      tags:
      - rules
swagger: "2.0"